server/pyproject.toml
[project]
name = "agentburg-server"
version = "0.1.0"
description = "AgentBurg world server"
requires-python = ">=3.13"
dependencies = [
    "agentburg-shared",
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyjwt[crypto]>=2.10.0",
    "argon2-cffi>=25.1.0",
    "nats-py>=2.9.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=1.0",
    "pytest-cov>=6.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/agentburg_server"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "S", "UP", "SIM"]
ignore = [
    "B008",   # Depends() in defaults — standard FastAPI pattern
    "S105",   # Hardcoded password — false positives on token_type
    "S104",   # Binding to 0.0.0.0 — required for Docker containers
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106", "S108", "S110"]

[dependency-groups]
dev = [
    "httpx-ws>=0.8.2",
]
